@@ -3,18 +3,22 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@reduxjs/toolkit": "^1.8.0",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.3",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.4.1",
     "@types/node": "^16.11.26",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "5.0.0",
-    "typescript": "^4.5.5",
-    "web-vitals": "^2.1.4"
+    "@types/react": "~17.0.39",
+    "@types/react-dom": "~17.0.11",
+    "@types/react-redux": "^7.1.22",
+    "formik": "~2.2.9",
+    "react": "~18.0.0",
+    "react-dom": "~18.0.0",
+    "react-redux": "~7.2.8",
+    "react-scripts": "~5.0.0",
+    "typescript": "~4.5.5",
+    "web-vitals": "~2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -39,5 +43,10 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "autoprefixer": "^10.4.2",
+    "postcss": "^8.4.7",
+    "tailwindcss": "~3.0.23"
   }
 }
